+++ telegram-bot-bumblebee-core/src/main/java/com/github/bumblebee/command/currency/service/CurrencyBidEvalService.java
@@ -1,10 +1,5 @@
 package com.github.bumblebee.command.currency.service;
 
-import com.github.bumblebee.command.currency.dao.CurrencyBidRepository;
-import com.github.bumblebee.command.currency.domain.CurrencyBid;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDate;
@@ -13,6 +8,12 @@ import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.github.bumblebee.command.currency.dao.CurrencyBidRepository;
+import com.github.bumblebee.command.currency.domain.CurrencyBid;
 
 @Service
 public class CurrencyBidEvalService {
@@ -24,7 +25,7 @@ public class CurrencyBidEvalService {
         this.repository = repository;
     }
 
-    public List<CurrencyBid> getTodayActualBids(Integer chatId) {
+    public List<CurrencyBid> getTodayActualBids(Long chatId) {
         Instant from = LocalDate.now().atStartOfDay(ZoneId.systemDefault()).toInstant();
         Instant to = from.plus(Duration.ofDays(1));
